fix: treat malformed gitfiles as unusable and split real-gateway tests
Recognize Git's invalid gitfile format as a local-integrity failure and
reject non-gitdir .git files in the local probe so prepare can clear and
recreate retained broken metadata. Split the real-gateway lifecycle suite
under the 300-line test-file threshold and add recovery coverage.

Generated-By: looper 0.11.0 (runner=fixer, agent=grok-build)
